requirements.txt
# Oracle Database connectivity
oracledb
cx_Oracle

# Data processing and analysis
pandas
numpy
sqlalchemy
openpyxl

# Web interface
streamlit

# CLI tools for database interaction
click
rich
tabulate

# Configuration and environment
python-dotenv
pyyaml

# Authentication
bcrypt>=4.0.0

# Logging and utilities
colorama
tqdm

# Confluence integration (спутниковый библиотекарь, полная индексация вложений)
requests
beautifulsoup4
pymupdf
python-docx
Pillow
pytesseract

# RAG and Vector Database
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# Спутниковый библиотекарь — агент на Gemini (инженер спутникового сегмента, куратор KB)
google-genai>=1.0.0

